chore(jobs): rename stream.function package to function
